Using cdn.rawgit.com instead of rawgit.com
This is the way RawGit wants people to use it. cdn.rawgit.com uses a permanent cache, so we link to a specific commit of the checklist. We'll have to update the link in this README when we want to link to new commits of the checklist. But this way we're playing nice with RawGit.